requirements.txt
gradio
torch
torchvision
transformers
opencv-python
ultralytics
Pillow
numpy 
streamlit 